Допиливая фабрику + меняю в HomeFragment.kt на новую. Допиливаю поиск. Делаю поиск по id и по имени. ставлю навигацию у activity_user.xml, пока с 1 фрагментом

+++ app/src/main/java/com/example/galleryapp/ui/fragments/home/SearchFragment.kt
@@ -16,7 +16,7 @@ class SearchFragment() :
         super.setObservers()
 
         viewModel.listener.observe(this) {
-            viewModel.searchPhotos(it.toInt())
+            viewModel.searchPhotos(it)
         }
     }
